rlp: tests: create invalid test file
